+++ httprunner/plugins/locusts/__main__.py
@@ -1,0 +1,6 @@
+import sys
+
+from httprunner.plugins.locusts import main
+
+if __name__ == "__main__":
+    main()
